ありがとうございます。

+++ src/app/admin/Menus/TitleCard.vue
@@ -1,34 +1,52 @@
 <template>
-  <div class="card block">
-    <div class="card-content">
-      <div class="media">
-        <div class="media-content" @click="toEdit()">
-          <h2 class="p-big" if v-if="title.name==''">
-            {{$t('editTitle.empty')}}...
-          </h2>
-          <h2 class="p-big bold" if v-else>
-            {{ title.name }}
-          </h2>
+  <div>
+    <!-- Title Card -->
+    <div class="bg-form r-8 p-l-16 p-r-16 p-t-16 p-b-16 m-t-24" @click="toEdit()">
+      <div class="t-h6 c-text-black-disabled" if v-if="title.name==''">{{$t('editTitle.empty')}}</div>
+      <div class="t-h6 c-text-black-disabled" if v-else>{{ title.name }}</div>
+    </div>
+
+    <!-- Card Actions -->
+    <div class="m-t-8 p-b-8 p-l-8 p-r-8">
+      <div class="cols">
+        <div class="flex-1">
+          <!-- Position Up -->
+          <b-button
+            class="b-reset op-button-pill h-36 bg-primary-bg m-r-8"
+            v-if="position!=='first'"
+            @click="positionUp"
+          >
+            <i class="material-icons c-primary s-18 p-l-8 p-r-8">arrow_upward</i>
+          </b-button>
+          <!-- Disable if First -->
+          <b-button class="b-reset op-button-pill h-36 bg-primary-bg m-r-8" disabled v-else>
+            <i class="material-icons c-primary s-18 p-l-8 p-r-8">arrow_upward</i>
+          </b-button>
+
+          <!-- Position Down -->
+          <b-button
+            class="b-reset op-button-pill h-36 bg-primary-bg m-r-8"
+            v-if="position!=='last'"
+            @click="positionDown"
+          >
+            <i class="material-icons c-primary s-18 p-l-8 p-r-8">arrow_downward</i>
+          </b-button>
+          <!-- Disable if Last -->
+          <b-button class="b-reset op-button-pill h-36 bg-primary-bg m-r-8" disabled v-else>
+            <i class="material-icons c-primary s-18 p-l-8 p-r-8">arrow_downward</i>
+          </b-button>
+
+          <!-- Duplicate -->
+          <b-button class="b-reset op-button-pill h-36 bg-primary-bg m-r-8" @click="forkItem">
+            <i class="material-icons c-primary s-18 p-l-8 p-r-8">queue</i>
+          </b-button>
+        </div>
+        <div>
+          <b-button class="b-reset op-button-pill h-36 bg-status-red-bg" @click="deleteItem">
+            <i class="material-icons c-status-red s-18 p-l-8 p-r-8">delete</i>
+          </b-button>
         </div>
       </div>
-    </div>
-    <div class="card-footer">
-      <a @click.prevent href="#" class="card-footer-item" @click="positionUp" v-if="position!=='first'">
-        <b-icon icon="arrow-up" size="is-midium"></b-icon>
-      </a>
-      <a @click.prevent class="card-footer-item" v-else>
-      </a>
-      <a @click.prevent href="#" class="card-footer-item" @click="positionDown" v-if="position!=='last'">
-        <b-icon icon="arrow-down" size="is-midium"></b-icon>
-      </a>
-      <a @click.prevent href="#" class="card-footer-item" v-else>
-      </a>
-      <a @click.prevent href="#" class="card-footer-item" @click="forkItem">
-        <b-icon icon="plus" size="is-midium"></b-icon>
-      </a>
-      <a @click.prevent href="#" class="card-footer-item" @click="deleteItem">
-        <b-icon icon="delete" size="is-midium"></b-icon>
-      </a>
     </div>
   </div>
 </template>
@@ -64,13 +82,11 @@ export default {
       this.$emit("forkItem", this.title.id);
     },
     deleteItem() {
-      this.$emit("deleteItem", this.title.id);
-    },
+      // this.$emit("deleteItem", this.title.id);
+      if (confirm(this.$t("editMenu.reallyDelete"))) {
+        this.$emit("deleteItem", this.title.id);
+      }
+    }
   }
 };
 </script>
-<style lang="scss" scoped>
-.card {
-  margin-bottom: 0.6rem;
-}
-</style>
